Continue work on enums
