+++ Tests/Cosmos.TestRunner.Full/TestKernelSets.cs
@@ -16,7 +16,7 @@ namespace Cosmos.TestRunner.Full
         public static IEnumerable<Type> GetStableKernelTypes()
         {
             //yield return typeof(BoxingTests.Kernel);
-            yield return typeof(Compiler.Tests.TypeSystem.Kernel);
+            //yield return typeof(Compiler.Tests.TypeSystem.Kernel);
             //yield return typeof(Compiler.Tests.Bcl.Kernel);
             //yield return typeof(Compiler.Tests.Bcl.System.Kernel);
             //yield return typeof(Cosmos.Compiler.Tests.Encryption.Kernel);
